Implement functionality to show user detail information

+++ app/src/main/java/com/zakir/androidtesting/repository/LocalUserDataSourceImp.java
@@ -20,7 +20,7 @@ public class LocalUserDataSourceImp implements UserDataSource {
     }
 
     @Override
-    public Flowable<User> getUserById(int userId) {
+    public Flowable<User> getUserById(long userId) {
         return userDao.getUser(userId);
     }
